fix: updated ui for room info modal, updated types
