+++ src/slices/hatch/booking-page/components/AvailableRoomCard.tsx
@@ -1,15 +1,14 @@
 'use client';
 import { Button, useDisclosure } from '@nextui-org/react';
-import { Check, InfoIcon, Plug, User, X } from 'lucide-react';
+import { InfoIcon, Plug, User } from 'lucide-react';
 import React from 'react';
 import { toast } from 'sonner';
 
-import { THatchRoom } from '@/constant/hatch-bookings/rooms-data';
 import { useSessionContext } from '@/slices/auth/context/SessionContext';
 import ConfirmationPopover from '@/slices/hatch/booking-page/components/ConfirmationPopover';
-import ResourcesIcon from '@/slices/hatch/booking-page/components/ResourcesIcon';
 import RoomInfoModal from '@/slices/hatch/booking-page/components/RoomInfoModal';
 import { useTimePickerContext } from '@/slices/hatch/booking-page/context/TimePickerContext';
+import { THatchRoom } from '@/slices/hatch/booking-page/types';
 
 type AvailableRoomCardProps = {
   roomInfo: THatchRoom;
@@ -18,7 +17,6 @@ type AvailableRoomCardProps = {
 export const AvailableRoomCard = ({ roomInfo }: AvailableRoomCardProps) => {
   const { handleAddBookRoom } = useTimePickerContext();
   const { isOpen, onOpen, onOpenChange } = useDisclosure();
-  const resourceKeys = Object.keys(roomInfo.resources);
   const { profile } = useSessionContext();
 
   function handleConfirmBookingWithMessage() {
@@ -38,20 +36,22 @@ export const AvailableRoomCard = ({ roomInfo }: AvailableRoomCardProps) => {
       >
         {roomInfo.roomName}
       </Button>
-      <div className='box-border hidden h-auto w-full flex-col items-center justify-between rounded-xl  border-2 bg-primary-800 p-4 text-center text-white md:flex'>
-        <div className='relative inline-block w-full items-center justify-center text-center'>
-          <div className='inline-block items-center font-bold'>
-            {roomInfo.roomName}
-          </div>
+
+      <div className='relative box-border hidden h-auto w-full flex-row items-center justify-between rounded-xl  border-2 bg-primary-800 p-4 text-center text-white md:flex'>
+        <div>
           <button
-            className='absolute right-0 top-0.5 rounded-full hover:bg-primary-700'
+            className='absolute left-2 top-2 rounded-full hover:bg-primary-700'
             onClick={onOpen}
           >
             <InfoIcon size={20} />
           </button>
         </div>
 
-        <div className='gap-x-4 my-2 mb-4 flex flex-row text-left'>
+        <div className='flex ml-4 items-center justify-center font-bold text-xl'>
+          {roomInfo.roomName}
+        </div>
+
+        <div className='gap-y-2 mx-4 flex flex-col text-left items-center justify-center'>
           <div className='flex items-center justify-center'>
             <User className='mr-1 w-[20px]' />
 
@@ -62,26 +62,16 @@ export const AvailableRoomCard = ({ roomInfo }: AvailableRoomCardProps) => {
             <Plug className='mr-1 w-[20px]' />
             <p>{roomInfo.outlets}</p>
           </div>
-
-          {resourceKeys.map((resource, index) => (
-            <div className='flex items-center justify-center' key={index}>
-              <ResourcesIcon resource={resource} />
-              {roomInfo.resources[resource] ? (
-                <Check className='w-[20px]' />
-              ) : (
-                <X className='w-[20px]' />
-              )}
-            </div>
-          ))}
         </div>
+
         <ConfirmationPopover
           handleConfirmBookingWithMessage={handleConfirmBookingWithMessage}
         >
           <Button
-            className='m-1 w-[130px] bg-white font-semibold hover:bg-gray'
+            className='m-1 w-[90px] border bg-inherit text-white text-md border-white font-semibold hover:bg-white hover:text-primary-700'
             size='sm'
           >
-            Book Now
+            Book
           </Button>
         </ConfirmationPopover>
       </div>
